+++ music/src/pages/login/login.css
@@ -1,0 +1,261 @@
+* {
+    margin: 0;
+    padding: 0;
+    -webkit-box-sizing: border-box;
+    box-sizing: border-box;
+  }
+  
+  *:before,
+  *:after {
+    -webkit-box-sizing: border-box;
+    box-sizing: border-box;
+  }
+  
+  a,
+  a:visited {
+    text-decoration: none;
+    font-family: 'StratosSkyeng', sans-serif;
+    cursor: pointer;
+  }
+  
+  button,
+  ._btn {
+    cursor: pointer;
+  }
+  
+  ul li {
+    list-style: none;
+  }
+  
+  /* @font-face {
+    font-family: 'StratosSkyeng';
+    src:
+      local('StratosSkyeng'),
+      local('StratosSkyeng'),
+      url('../fonts/StratosSkyeng.woff2') format('woff2'),
+      url('../fonts/StratosSkyeng.woff') format('woff'),
+      url('../fonts/StratosSkyeng.ttf') format('truetype');
+    font-weight: 400;
+    font-style: normal;
+  } */
+  
+  @font-face {
+    font-family: 'StratosSkyeng';
+    src:
+      local('StratosSkyeng'),
+      local('StratosSkyeng'),
+      url('../../../public/fonts/StratosSkyeng.woff2') format('woff2'),
+      url('../../../public/fonts/StratosSkyeng.woff') format('woff');
+    font-weight: 400;
+    font-style: normal;
+  }
+  
+  html,
+  body {
+    width: 100%;
+    height: 100%;
+    font-family: 'StratosSkyeng', sans-serif;
+  }
+  
+  .wrapper {
+    width: 100%;
+    min-height: 100%;
+    overflow: hidden;
+  }
+  
+  .container-enter {
+    max-width: 100%;
+    height: 100vh;
+    margin: 0 auto;
+    position: relative;
+    background-color: rgba(0, 0, 0, 0.85);
+  }
+  
+  .modal__btn-enter:hover {
+    background-color: #3f007d;
+  }
+  
+  .modal__btn-signup:hover {
+    background-color: #f4f5f6;
+  }
+  
+  .modal__btn-enter:active {
+    background-color: #271a58;
+  }
+  
+  .modal__btn-signup:active {
+    background-color: #d9d9d9;
+  }
+  
+  .modal__block {
+    position: absolute;
+    z-index: 2;
+    left: calc(50% - (366px / 2));
+    top: calc(50% - (439px / 2));
+    opacity: 1;
+  }
+  
+  .modal__form-login {
+    width: 366px;
+    height: 439px;
+    background-color: #ffffff;
+    border-radius: 12px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: column;
+    flex-direction: column;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    padding: 43px 47px 47px 40px;
+  }
+  
+  .modal__form-login input:first-child {
+    margin-bottom: 30px;
+  }
+  
+  .modal__logo {
+    width: 140px;
+    height: 21px;
+    margin-bottom: 34px;
+    background-color: transparent;
+  }
+  
+  .modal__logo img {
+    width: 140px;
+    height: auto;
+  }
+  
+  .modal__input {
+    width: 100%;
+    border-top: none;
+    border-left: none;
+    border-right: none;
+    border-bottom: 1px solid #d0cece;
+    padding: 8px 1px;
+  }
+  
+  .modal__input::-webkit-input-placeholder {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+    letter-spacing: -0.05px;
+    color: #d0cece;
+  }
+  
+  .modal__input:-ms-input-placeholder {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+    letter-spacing: -0.05px;
+    color: #d0cece;
+  }
+  
+  .modal__input::-ms-input-placeholder {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+    letter-spacing: -0.05px;
+    color: #d0cece;
+  }
+  
+  .modal__input::placeholder {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+    letter-spacing: -0.05px;
+    color: #d0cece;
+  }
+  
+  .modal__btn-enter {
+    width: 278px;
+    height: 52px;
+    background-color: #580ea2;
+    border-radius: 6px;
+    margin-top: 60px;
+    margin-bottom: 20px;
+    border: none;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+  }
+  
+  .modal__btn-enter a {
+    width: 100%;
+    height: 100%;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+    letter-spacing: -0.05px;
+    color: #ffffff;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+  }
+  
+  .modal__btn-signup {
+    width: 278px;
+    height: 52px;
+    background-color: transparent;
+    border: 1px solid #d0cece;
+    border-radius: 6px;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+    letter-spacing: -0.05px;
+    color: #000000;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+  }
+  
+  .modal__btn-signup a {
+    width: 100%;
+    height: 100%;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 24px;
+    letter-spacing: -0.05px;
+    color: #000000;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+  }
+  
+  .login {
+    margin-bottom: 30px;
+  }
