+++ music/src/routes.jsx
@@ -3,7 +3,7 @@ import { MainPage } from "./pages/main";
 import { CategoryPage } from "./pages/category";
 import { FavoritesPage } from "./pages/favorites";
 import { NotFoundPage } from "./pages/not-found";
-import { LoginPage } from "./pages/login";
+import { LoginPage } from "./pages/login/login";
 import { RegisterPage } from "./pages/register";
 import { ProtectedRoute } from './components/protected-route';
